requirements.txt
spade==4.0.1
aiohttp>=3.8.4
beautifulsoup4==4.12.2
google-generativeai==0.3.1
python-dotenv==1.0.0
pandas==2.1.4
pyjwt==2.8.0
redis==5.0.1
SQLAlchemy==2.0.25
aiosqlite==0.19.0
pytest==7.4.3
pytest-asyncio==0.23.2
requests==2.31.0
python-jose==3.3.0
fastapi==0.109.0
uvicorn==0.25.0
